edited the code

+++ inc/CANService.h
@@ -48,8 +48,8 @@ int CAN_Init(CAN_CHANNEL_E can_ch);
 int CAN_DeInit(CAN_CHANNEL_E can_ch);
 int CAN_SetBaudrate(CAN_CHANNEL_E can_ch, int baud);
 int CAN_SetMode(CAN_CHANNEL_E can_ch, uint8_t mode);
-int CAN_SendPacket(CAN_CHANNEL_E can_ch, uint32_t can_id, uint8_t can_dlc, uint8_t* data);
-int CAN_RecvPacket(CAN_CHANNEL_E can_ch, uint32_t* can_id, uint8_t* can_dlc, uint8_t* data);
+int CAN_SendPacket(CAN_CHANNEL_E can_ch, uint16_t can_id, uint8_t can_dlc, uint8_t* data);
+int CAN_RecvPacket(CAN_CHANNEL_E can_ch, uint16_t* can_id, uint8_t* can_dlc, uint8_t* data);
 
 int CAN_Rwlock_Init(void);
 int CAN_Rwlock_DeInit(void);
